requirements.txt
streamlit
numpy
opencv-python
torch
torchvision
matplotlib
scikit-image
Pillow